+++ services/buyer_frontend_service/buyer_frontend_main.cc
@@ -40,6 +40,7 @@
 #include "services/common/encryption/key_fetcher_factory.h"
 #include "services/common/metric/server_definition.h"
 #include "services/common/telemetry/configure_telemetry.h"
+#include "services/common/util/signal_handler.h"
 #include "services/common/util/status_macros.h"
 #include "src/core/lib/event_engine/default_event_engine.h"
 #include "src/cpp/concurrent/event_engine_executor.h"
@@ -54,6 +55,10 @@ ABSL_FLAG(std::optional<std::string>, buyer_kv_server_addr, std::nullopt,
 // Added for performance/benchmark testing of both types of http clients.
 ABSL_FLAG(std::optional<int>, generate_bid_timeout_ms, std::nullopt,
           "Max time to wait for generate bid request to finish.");
+ABSL_FLAG(std::optional<int>, protected_app_signals_generate_bid_timeout_ms,
+          std::nullopt,
+          "Max time to wait for protected app signals generate bid request to "
+          "finish.");
 ABSL_FLAG(std::optional<int>, bidding_signals_load_timeout_ms, std::nullopt,
           "Max time to wait for fetching bidding signals to finish.");
 ABSL_FLAG(std::optional<bool>, enable_buyer_frontend_benchmarking, std::nullopt,
@@ -90,6 +95,8 @@ absl::StatusOr<TrustedServersConfigClient> GetConfigClient(
   config_client.SetFlag(FLAGS_bidding_server_addr, BIDDING_SERVER_ADDR);
   config_client.SetFlag(FLAGS_buyer_kv_server_addr, BUYER_KV_SERVER_ADDR);
   config_client.SetFlag(FLAGS_generate_bid_timeout_ms, GENERATE_BID_TIMEOUT_MS);
+  config_client.SetFlag(FLAGS_protected_app_signals_generate_bid_timeout_ms,
+                        PROTECTED_APP_SIGNALS_GENERATE_BID_TIMEOUT_MS);
   config_client.SetFlag(FLAGS_bidding_signals_load_timeout_ms,
                         BIDDING_SIGNALS_LOAD_TIMEOUT_MS);
   config_client.SetFlag(FLAGS_enable_buyer_frontend_benchmarking,
@@ -216,7 +223,13 @@ absl::Status RunServer() {
       GetBidsConfig{
           config_client.GetIntParameter(GENERATE_BID_TIMEOUT_MS),
           config_client.GetIntParameter(BIDDING_SIGNALS_LOAD_TIMEOUT_MS),
-          config_client.GetBooleanParameter(ENABLE_ENCRYPTION)},
+          config_client.GetBooleanParameter(ENABLE_ENCRYPTION),
+          config_client.GetIntParameter(
+              PROTECTED_APP_SIGNALS_GENERATE_BID_TIMEOUT_MS),
+          config_client.GetBooleanParameter(ENABLE_PROTECTED_APP_SIGNALS),
+          config_client.GetBooleanParameter(ENABLE_OTEL_BASED_LOGGING),
+          std::string(config_client.GetStringParameter(CONSENTED_DEBUG_TOKEN)),
+      },
       enable_buyer_frontend_benchmarking);
 
   grpc::EnableDefaultHealthCheckService(true);
@@ -261,6 +274,7 @@ absl::Status RunServer() {
 }  // namespace privacy_sandbox::bidding_auction_servers
 
 int main(int argc, char** argv) {
+  signal(SIGSEGV, privacy_sandbox::bidding_auction_servers::SignalHandler);
   absl::ParseCommandLine(argc, argv);
   google::InitGoogleLogging(argv[0]);