+++ services/common/telemetry/telemetry_flag.h
@@ -20,6 +20,7 @@
 #include <string>
 #include <utility>
 
+#include "absl/container/flat_hash_map.h"
 #include "absl/status/statusor.h"
 #include "absl/strings/string_view.h"
 #include "services/common/metric/definition.h"